+++ src/V1/Assurances/EnsureAllOf.php
@@ -45,13 +45,17 @@ namespace GanbaroDigital\Defensive\V1\Assurances;
 
 use GanbaroDigital\Defensive\V1\Exceptions\DefensiveExceptions;
 use GanbaroDigital\Defensive\V1\Interfaces\Assurance;
+use GanbaroDigital\Defensive\V1\Interfaces\ListAssurance;
 use GanbaroDigital\Defensive\V1\Requirements\RequireValidAssurances;
 use GanbaroDigital\DIContainers\V1\Interfaces\FactoryList;
 
-class EnsureAllOf implements Assurance
+class EnsureAllOf implements Assurance, ListAssurance
 {
     // saves us having to declare ::__invoke() ourselves
     use InvokeableAssurance;
+
+    // saves us having to declare ::toList() ourselves
+    use ListableAssurance;
 
     /**
      * the exceptions we should use