+++ docs/changelog.md
@@ -18,6 +18,7 @@ pageflow_next_text: Contributing
   - added `ListRequirement` interface
   - added `ListableAssurance` trait
   - `ComposableAssurance` now implements `ListAssurance` interface
+  - `EnsureAllOf` now implements `ListAssurance` interface
 
 ## v1.2016062801